Move saving and discarding into publish box
