+++ src/app/story/directives/hero.component.spec.ts
@@ -1,4 +1,4 @@
-import { cit, create, provide, stubPipe, niceEl, By } from '../../../testing';
+import { cit, create, provide, stubPipe } from '../../../testing';
 import { Router } from '@angular/router';
 import { RouterStub } from '../../../testing/stub.router';
 import { ToastrService } from 'ngx-prx-styleguide';
@@ -7,30 +7,10 @@ import { StoryHeroComponent } from './hero.component';
 describe('StoryHeroComponent', () => {
 
   create(StoryHeroComponent);
-
   provide(Router, RouterStub);
   provide(ToastrService, {success: () => {}});
 
   stubPipe('timeago');
-
-  const expectDisabled = (el, text, shouldBeDisabled) => {
-    let button = el.queryAll(By.css('prx-button')).find(btn => {
-      return btn.nativeElement.textContent.trim() === text;
-    });
-    if (!button) {
-      fail(`Could not find button with text: ${text}`);
-    } else {
-      let isDisabled = button.nativeElement.disabled;
-      let ngDisabled = button.nativeElement.getAttribute('ng-reflect-disabled');
-      isDisabled = isDisabled || ngDisabled;
-      if (shouldBeDisabled && isDisabled === null) {
-        fail(`Expected disabled - ${niceEl(button)}`);
-      }
-      if (!shouldBeDisabled && isDisabled !== null) {
-        fail(`Expected enabled - ${niceEl(button)}`);
-      }
-    }
-  };
 
   cit('shows edit vs create state', (fix, el, comp) => {
     expect(el).toContainText('Create Episode');
@@ -38,38 +18,4 @@ describe('StoryHeroComponent', () => {
     fix.detectChanges();
     expect(el).toContainText('Edit Episode');
   });
-
-  cit('unstrictly saves new stories', (fix, el, comp) => {
-    comp.story = {isNew: true, changed: () => null,  invalid: () => 'bad'};
-    fix.detectChanges();
-    expect(el).toContainText('Save');
-    expectDisabled(el, 'Save', true);
-    comp.story.invalid = (f, strict) => strict ? 'bad' : null;
-    fix.detectChanges();
-    expectDisabled(el, 'Save', false);
-  });
-
-  cit('unstrictly saves unpublished stories', (fix, el, comp) => {
-    comp.story = {changed: () => null, invalid: () => 'bad'};
-    fix.detectChanges();
-    expect(el).toContainText('Save');
-    expectDisabled(el, 'Save', true);
-    comp.story.invalid = (f, strict) => strict ? 'bad' : null;
-    fix.detectChanges();
-    expectDisabled(el, 'Save', false);
-  });
-
-  cit('strictly saves published stories', (fix, el, comp) => {
-    comp.story = {publishedAt: new Date(), isPublished: () => null, changed: () => null, invalid: () => 'bad'};
-    fix.detectChanges();
-    expect(el).toContainText('Save');
-    expectDisabled(el, 'Save', true);
-    comp.story.invalid = (f, strict) => strict ? 'bad' : null;
-    fix.detectChanges();
-    expectDisabled(el, 'Save', true);
-    comp.story.invalid = (f, strict) => null;
-    fix.detectChanges();
-    expectDisabled(el, 'Save', false);
-  });
-
 });
