+++ src/app/story/directives/status.component.spec.ts
@@ -1,4 +1,4 @@
-import { cit, create, provide, stubPipe, By } from '../../../testing';
+import { cit, create, provide, stubPipe, niceEl, By } from '../../../testing';
 import { Router } from '@angular/router';
 import { Angulartics2 } from 'angulartics2';
 import { RouterStub } from '../../../testing/stub.router';
@@ -24,6 +24,25 @@ describe('StoryStatusComponent', () => {
   provide(ToastrService);
   provide(Angulartics2, {trackLocation: () => {}});
 
+  const expectDisabled = (el, text, shouldBeDisabled) => {
+    let button = el.queryAll(By.css('prx-button')).find(btn => {
+      return btn.nativeElement.textContent.trim() === text;
+    });
+    if (!button) {
+      fail(`Could not find button with text: ${text}`);
+    } else {
+      let isDisabled = button.nativeElement.disabled;
+      let ngDisabled = button.nativeElement.getAttribute('ng-reflect-disabled');
+      isDisabled = isDisabled || ngDisabled;
+      if (shouldBeDisabled && isDisabled === null) {
+        fail(`Expected disabled - ${niceEl(button)}`);
+      }
+      if (!shouldBeDisabled && isDisabled !== null) {
+        fail(`Expected enabled - ${niceEl(button)}`);
+      }
+    }
+  };
+
   const mockStory = (story, comp, fix) => {
     comp.story = story;
     comp.story.invalid = comp.story.invalid || (() => null);
@@ -32,6 +51,39 @@ describe('StoryStatusComponent', () => {
     comp.id = comp.story.isNew ? null : 1234;
     fix.detectChanges();
   };
+
+  cit('unstrictly saves new stories', (fix, el, comp) => {
+    mockStory({isNew: true, invalid: () => 'bad'}, comp, fix);
+    fix.detectChanges();
+    expect(el).toContainText('Save');
+    expectDisabled(el, 'Save', true);
+    comp.story.invalid = (f, strict) => strict ? 'bad' : null;
+    fix.detectChanges();
+    expectDisabled(el, 'Save', false);
+  });
+
+  cit('unstrictly saves unpublished stories', (fix, el, comp) => {
+    mockStory({invalid: () => 'bad'}, comp, fix);
+    fix.detectChanges();
+    expect(el).toContainText('Save');
+    expectDisabled(el, 'Save', true);
+    comp.story.invalid = (f, strict) => strict ? 'bad' : null;
+    fix.detectChanges();
+    expectDisabled(el, 'Save', false);
+  });
+
+  cit('strictly saves published stories', (fix, el, comp) => {
+    mockStory({publishedAt: new Date(), isPublished: () => null, invalid: () => 'bad'}, comp, fix);
+    fix.detectChanges();
+    expect(el).toContainText('Save');
+    expectDisabled(el, 'Save', true);
+    comp.story.invalid = (f, strict) => strict ? 'bad' : null;
+    fix.detectChanges();
+    expectDisabled(el, 'Save', true);
+    comp.story.invalid = (f, strict) => null;
+    fix.detectChanges();
+    expectDisabled(el, 'Save', false);
+  });
 
   cit('shows story status', (fix, el, comp) => {
     mockStory({isNew: true}, comp, fix);
